refactor: share FullscreenPassScope (VR/PerfMode)
Promote PerfMode's private FullscreenPassScope RAII guard to
Util::FullscreenPassScope (Utils/D3D.h + .cpp) and add PS constant-
buffer slot 1 to its saved set. Replace the two hand-rolled save/restore
blocks in VRStereoOptimizations (ExecuteStencilWritePass,
ExecuteDepthFillPass) with it.

DispatchGBufferFill is intentionally left as-is: it is a compute pass
that only needs RTV/DSV save/restore, so the full graphics-state scope
would be a wrong-fit superset.

Behavior-preserving: the scope is the already-shipping PerfMode helper
(unchanged but for the additive PSCB1 field, a latent leak fix), and the
VR conversions are a superset restore of the old inline blocks. Builds
clean (ALL preset). Net -138 lines.

Closes #144.
